Fixed several issues exposed by the Microsoft compiler.
git-svn-id: http://voip.null.ro/svn/yate@6459 acf43c95-373e-0410-b603-e72c3f656dc1

+++ engine/String.cpp
@@ -327,7 +327,7 @@ bool UChar::decode(DataBlock& buff, Endianness order, uint32_t maxChar)
     len = len >> 1;
     if (!decode(in,len,order,maxChar))
 	return false;
-    buff.cut(-(buff.length() - len * 2));
+    buff.cut(-(int)(buff.length() - len * 2));
     return true;
 }
